+++ django_echarts/entities/__init__.py
@@ -5,7 +5,7 @@ from .base import DwString
 from .chart_widgets import NamedCharts
 from .containers import RowContainer, Container
 from .html_widgets import (
-    HTMLBase, LinkItem, Menu, Jumbotron, Nav, Copyright, Message, ValuesPanel, ValueItem
+    HTMLBase, LinkItem, Menu, Jumbotron, Nav, Copyright, Message, ValuesPanel, ValueItem, Title
 )
 from .pages import (WidgetCollection, WidgetGetterMixin)
 from .styles import (bootstrap_table_class, material_table_class)